backend/requirements.txt
fastapi
uvicorn
pymongo
pathway
langchain-groq
litellm
tavily-python
langchain-community
python-jose[cryptography]
passlib[bcrypt]
httpx
sendgrid